+++ inc/hcd.h
@@ -17,10 +17,10 @@
  * Host Controller Endpoint Descriptor, refer to Section 4.2, Endpoint Descriptor
  */
 typedef struct _HC_ED {
-	volatile ulong	Control;	/* dword 0*/
-	volatile ulong	TailP;		/* physical pointer to HC_TD*/
-	volatile ulong	HeadP;		/* flags + phys ptr to HC_TD*/
-	volatile ulong	NextED;		/* phys ptr to HC_ED*/
+	volatile uint32_t	Control;	/* dword 0*/
+	volatile uint32_t	TailP;		/* physical pointer to HC_TD*/
+	volatile uint32_t	HeadP;		/* flags + phys ptr to HC_TD*/
+	volatile uint32_t	NextED;		/* phys ptr to HC_ED*/
 } HC_ED, *PHC_ED;
 
 #define HcED_FA				0x0000007F	/* */
@@ -47,9 +47,9 @@ typedef struct _HC_ED {
  * Host Controller Transfer Descriptor, refer to Section 4.3, Transfer Descriptor
  */
 typedef struct _HC_TD {
-	volatile ulong	Control;	/* dword 0*/
+	volatile uint32_t	Control;	/* dword 0*/
 	volatile void *	CBP;		/* phys ptr to data buffer*/
-	volatile ulong	NextTD;		/* phys ptr to HC_TD*/
+	volatile uint32_t	NextTD;		/* phys ptr to HC_TD*/
 	volatile void *	BE;			/* phys ptr to end of data buffer*/
 } HC_TD, *PHC_TD;
 
@@ -88,10 +88,10 @@ typedef struct _HC_TD {
 */
 typedef struct _HCD_ED {
 	HC_ED		HcED;
-	uchar		type;
-	uchar		interval;
-	uchar		flag;
-	uchar		reserved;
+	uint8_t		type;
+	uint8_t		interval;
+	uint8_t		flag;
+	uint8_t		reserved;
 	void		*dev;
 	LIST_ENTRY	ed_list;
 } HCD_ED, *PHCD_ED;
@@ -100,26 +100,26 @@ typedef struct _HCD_ED {
 * HCD Transfer Descriptor
 */
 typedef struct _HCD_TD {
-	HC_TD		HcTD;
-	PURB		urb;
-	int			index;
+	HC_TD   HcTD;
+	PURB    urb;
+	int32_t index;
 	LIST_ENTRY	list;
 } HCD_TD, *PHCD_TD;
 
 
 struct hcca{
-	ulong		int_table[32];	/* These 32 Dwords are pointers to interrupt EDs */
-	ushort		framenumber;	/* Contains the current frame number.  */
-	ushort		pad1;			/* When the HC updates HccaFrameNumber, it sets this word to 0. */
-	ulong		donehead;		/* The Done Queue of completed TD's. */
-	uchar		reserved[116];
+	uint32_t		int_table[32];	/* These 32 Dwords are pointers to interrupt EDs */
+	uint16_t		framenumber;	/* Contains the current frame number.  */
+	uint16_t		pad1;			/* When the HC updates HccaFrameNumber, it sets this word to 0. */
+	uint32_t		donehead;		/* The Done Queue of completed TD's. */
+	uint8_t		reserved[116];
 };
 
 struct hcd_info{
 	struct hcca *hcca;			/* address of HCCA */
 
-	int 		disabled;		/* status of host controller */
-	ulong		hc_control;		/* copy of the hc control reg */
+	int32_t disabled;		/* status of host controller */
+	uint32_t		hc_control;		/* copy of the hc control reg */
 
 	LIST_ENTRY	ed_bulk;		/* bulk endpoint list*/
 	LIST_ENTRY	ed_control;		/* control endpoint list*/
@@ -127,7 +127,7 @@ struct hcd_info{
 	LIST_ENTRY	ed_pause0;		/* delete endpoint list0*/
 	LIST_ENTRY	ed_pause1;		/* delete endpoint list1*/
 
-	long 		rh_status;		/* status of root hub */
+	int32_t rh_status;		/* status of root hub */
 	void *		rh_device;
 };
 
@@ -140,17 +140,17 @@ extern struct hcd_info hcd_info;
 #define disable_ohci_irq() put_wvalue(HostCtl, B_OHCIIRQ_MASK)
 #define enable_ohci_irq() put_wvalue(HostCtl, 0)
 
-int hcd_init(void);
+int32_t hcd_init(void);
 void hcd_exit(void);
 
-int hcd_rh_events(void);
+int32_t hcd_rh_events(void);
 
-PHCD_ED hcd_create_ed(uchar, uchar, uchar, uchar, ushort, uchar);
-int hcd_update_ed(PHCD_ED, uchar, ushort);
+PHCD_ED hcd_create_ed(uint8_t, uint8_t, uint8_t, uint8_t, uint16_t, uint8_t);
+int32_t hcd_update_ed(PHCD_ED, uint8_t, uint16_t);
 void hcd_delete_ed(PHCD_ED);
 
-int hcd_transfer_request(PURB);
-int hcd_transfer_cancel(PURB);
-int hcd_rh_disconnect(void);
+int32_t hcd_transfer_request(PURB);
+int32_t hcd_transfer_cancel(PURB);
+int32_t hcd_rh_disconnect(void);
 
 #endif	/* _HCD_H_ */
